Move WideColumn(s) to a public header
