+++ db/wide/wide_column_serialization.cc
@@ -9,6 +9,7 @@
 #include <cassert>
 #include <limits>
 
+#include "rocksdb/slice.h"
 #include "util/autovector.h"
 #include "util/coding.h"
